+++ utils/ref_optics.py
@@ -2,30 +2,24 @@
 
 nk_ref_dict = {}
 
-Gao_2016 = '''Gao, Y. , Le Corre, V. M., Gaï¿½tis, A. , Neophytou, M. , Hamid, M. A., Takanabe, K. and Beaujuge, P. M. 
-(2016), Homo-Tandem Polymer Solar Cells with VOC >1.8 V for Efficient PV-Driven Water Splitting. Adv. Mater., 28: 3366-3373. 
-doi:10.1002/adma.201504633'''
+Gao_2016 = '''
+Y. Gao, V.M. Le Corre, A. Gaïtis, M. Neophytou, M.A. Hamid, K. Takanabe, and P.M. Beaujuge, Homo-Tandem Polymer Solar Cells with VOC >1.8 V for Efficient PV-Driven Water Splitting. Adv. Mater. 28, 17 (2016), 3366-3373.
+'''
 nk_ref_dict['nk_ZnONP'] = Gao_2016
 nk_ref_dict['nk_aZnO'] = Gao_2016
 
 Liu_2005 = '''
-Z.T. Liu, C.Y. Kwong, C.H. Cheung, A.B. Djuriï¿½ic, Y. Chan, P.C. Chui,
-The characterization of the optical functions of BCP and CBP thin films by spectroscopic ellipsometry,
-Synthetic Metals, Volume 150, Issue 2, 2005, Pages 159-163,
-https://doi.org/10.1016/j.synthmet.2005.02.001.
+Z.T. Liu, C.Y. Kwong, C.H. Cheung, A.B. Djurïsic, Y. Chan, and P.C. Chui, The characterization of the optical functions of BCP and CBP thin films by spectroscopic ellipsometry, Synthetic Metals 150, 2 (2005), 159-163.
 '''
 nk_ref_dict['nk_BCPLiu'] = Liu_2005
 
 Sherkar_2017_1 = '''
-T. S. Sherkar, C. Momblona, L. Gil-Escrig, H. J. Bolink, L. J. A. Koster, Adv. Energy Mater. 2017, 7, 1602432. https://doi.org/10.1002/aenm.201602432
+T. S. Sherkar, C. Momblona, L. Gil-Escrig, H. J. Bolink, and L. J. A. Koster, Improving perovskite solar cells: insights from a validated device model, Adv. Energy Mater. 7, 13 (2017), 1602432.
 '''
 nk_ref_dict['nk_MAPI_2'] = Sherkar_2017_1
 
 Sherkar_2017_2 = '''
-Recombination in Perovskite Solar Cells: Significance of Grain Boundaries, Interface Traps, and Defect Ions
-Tejas S. Sherkar, Cristina Momblona, Lidï¿½n Gil-Escrig, Jorge ï¿½vila, Michele Sessolo, Henk J. Bolink, and L. Jan Anton Koster
-ACS Energy Letters 2017 2 (5), 1214-1222
-DOI: 10.1021/acsenergylett.7b00236
+T. S. Sherkar, C. Momblona, L. Gil-Escrig, H. J. Bolink, and L. J. A. Koster, Recombination in perovskite solar cells: significance of grain boundaries, interface traps, and defect ions, ACS Energy Lett. 2,5 (2017), 1214-1222.
 '''
 
 nk_ref_dict['nk_TaTm'] = Sherkar_2017_2
@@ -33,16 +27,13 @@ nk_ref_dict['nk_C60_2'] = Sherkar_2017_2
 nk_ref_dict['nk_MAPI'] = Sherkar_2017_2
 
 Qiu_2015 = '''
-An electron beam evaporated TiO2 layer for high efficiency planar perovskite solar cells on flexible polyethylene terephthalate substrates
-Weiming Qiu,*ab  Ulrich W. Paetzold,ad  Robert Gehlhaar,a  Vladimir Smirnov,d  Hans-Gerd Boyen,g  Jeffrey G. Tait,a  Bert Conings,g  Weimin Zhang,e  Christian B. Nielsen,e  Iain McCulloch,f  Ludo Froyen,b  Paul Heremans*ac  and  David Cheynsa
-J. Mater. Chem. A, 2015,3, 22824-22829 
-10.1039/C5TA07515G
+W. Qiu, U.W. Paetzold, R. Gehlhaar, V. Smirnov, H. Boyen, J.G. Tait, B. Conings, W. Zhang, C.B. Nielsen, I. McCulloch, L. Froyen, P. Heremans,  and  D. Cheynsa, An electron beam evaporated TiO2 layer for high efficiency planar perovskite solar cells on flexible polyethylene terephthalate substrates, J. Mater. Chem. A 3, 45 (2015), 22824-22829.
 '''
 
 nk_ref_dict['nk_TiO2'] = Qiu_2015
 
 NC_2020 = '''
-Nature Communications volume 11, Article number: 5220 (2020) https://doi.org/10.1038/s41467-020-19029-9
+Y. Firdaus, V.M. Le Corre, S. Karuthedath, W. Liu, A. Markina, W. Huang, S. Chattopadhyay, M.M. Nahid, M.I. Nugraha, Y. Lin, A. Seitkhan, A. Basu, W. Zhang, I. McCulloch, H. Ade, J. Labram, F. Laquai, D. Andrienko, L.J.A. Kostr, and T.D. Anthopoulos, Nature Comm. 11, 1 (2020), 5220.
 '''
 nk_ref_dict['nk_IDTBR'] = NC_2020
 nk_ref_dict['nk_IT2Cl'] = NC_2020
@@ -57,19 +48,19 @@ nk_ref_dict['nk_DPO'] = NC_2020
 nk_ref_dict['nk_CuSCN'] = NC_2020
 
 Corre_2019 = '''
-ACS Appl. Energy Mater. 2019, 2, 9, 6280–6287 https://doi.org/10.1021/acsaem.9b00856
+V.M. Le Corre, M. Stolterfoht, L. Toro Perdigon, M. Feuerstein , C. Wolff, L. Gil-Escrig, H.J. Bolink, D. Neher, and L.J.A. Koster, Charge transport layers limiting the efficiency of perovskite solar cells: how to optimize conductivity, doping, and thickness, ACS Appl. Energy Mater 2, 9 (2019), 6280-6287 
 '''
 
 nk_ref_dict['nk_C60_1'] = Corre_2019
 nk_ref_dict['nk_PTAA'] = Corre_2019
 
 Burkhard_2010 = '''
-G. F. Burkhard, E. T. Hoke, M. D. McGehee, Adv. Mater., 22, 3293. Accounting for Interference, Scattering, and Electrode Absorption to Make
-Accurate Internal Quantum Efficiency Measurements in Organic and Other Thin Solar Cells
+G. F. Burkhard, E. T. Hoke, and M. D. McGehee, Accounting for Interference, Scattering, and Electrode Absorption to Make Accurate Internal Quantum Efficiency Measurements in Organic and Other Thin Solar Cells, Adv. Mater. 22, 30 (2010), 3293-3297
 '''
 nk_ref_dict['nk_Ag'] = Burkhard_2010
 nk_ref_dict['nk_Air'] = Burkhard_2010
 nk_ref_dict['nk_Al'] = Burkhard_2010
+nk_ref_dict['nk_Au'] = Burkhard_2010
 nk_ref_dict['nk_Ca'] = Burkhard_2010
 nk_ref_dict['nk_ITO'] = Burkhard_2010
 nk_ref_dict['nk_P3HT'] = Burkhard_2010
@@ -81,18 +72,20 @@ nk_ref_dict['nk_SiO2'] = Burkhard_2010
 nk_ref_dict['nk_aSi'] = Burkhard_2010
 
 Xie_2017 = '''
-Xie, Ziang, et al. "Refractive index and extinction coefficient of NH2CH= NH2PbI3 perovskite photovoltaic material." Journal of Physics: Condensed Matter 29.24 (2017): 245702.
+Z. Xie, S. Sun, Y. Yan, L. Zhang, R. Hou, F. Ruixiang, and G.G. Qin, Refractive index and extinction coefficient of NH2CH= NH2PbI3 perovskite photovoltaic material, J. Phys.: Condens. Matter 29, 24 (2017), 245702.
 '''
 nk_ref_dict['FAPI'] = Xie_2017
 
 Werner_2018 = '''
-Werner, J., Nogay, G., Sahli, F., Yang, T. C. J., Bräuninger, M., Christmann, G., ... & Ballif, C. (2018). Complex refractive indices of cesium–formamidinium-based mixed-halide perovskites with optical band gaps from 1.5 to 1.8 eV. ACS Energy Letters, 3(3), 742-74
+J. Werner, G. Nogay, F. Sahli, T. Chien-Jen Yang, M. Brauninger, G. Christmann, A. Walter, B.A. Kamino, P. Fiala, P. Loper, S. Nicolay, Q. Jeangros, B. Niesen, and C. Balif, Complex refractive indices of cesium-formamidinium-based mixed-halide perovskites with optical band gaps from 1.5 to 1.8 eV, ACS Energy Lett. 3, 3 (2018), 742-747
 '''
 
 nk_ref_dict['nk_FACsPbIBr'] = Werner_2018
 
 spectrum_ref_dict = {}
 
-AM_REF = 'Gueymard, C. A., Myers, D., & Emery, K. (2002). Proposed reference irradiance spectra for solar energy systems testing. Solar energy, 73(6), 443-467 https://doi.org/10.1016/S0038-092X(03)00005-7'
+AM_REF = '''
+C.A. Gueymard, D. Myers, and K. Emery, Proposed reference irradiance spectra for solar energy systems testing, Solar energy 73, 6 (2002), 443-467
+'''
 
 spectrum_ref_dict['AM15G'] = AM_REF